[#171] Utils Class to Collect Answer from View, remove answer collection methods from activity
