+++ app/src/main/java/np/com/naxa/factsnepal/surveys/SurveyActivity.java
@@ -2,18 +2,10 @@ package np.com.naxa.factsnepal.surveys;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.text.TextUtils;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
-import android.widget.CheckBox;
-import android.widget.EditText;
 import android.widget.LinearLayout;
-import android.widget.RadioButton;
-import android.widget.RadioGroup;
-import android.widget.RatingBar;
-import android.widget.Spinner;
-import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -29,7 +21,6 @@ import java.io.EOFException;
 import java.net.SocketTimeoutException;
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 
 import io.reactivex.Observable;
 import io.reactivex.ObservableSource;
@@ -43,13 +34,12 @@ import np.com.naxa.factsnepal.common.BaseActivity;
 import np.com.naxa.factsnepal.common.Constant;
 import np.com.naxa.factsnepal.userprofile.UserLoginResponse;
 import np.com.naxa.factsnepal.utils.ActivityUtil;
+import np.com.naxa.factsnepal.utils.GetDataFromJsonView;
 import np.com.naxa.factsnepal.utils.JsonView;
 import np.com.naxa.factsnepal.utils.SharedPreferenceUtils;
 
-import static np.com.naxa.factsnepal.common.Constant.JsonKeySurvey.KEY_ANS_ID;
+import static np.com.naxa.factsnepal.common.Constant.JsonKeySurvey.KEY_ANSWER;
 import static np.com.naxa.factsnepal.common.Constant.JsonKeySurvey.KEY_COMPANY_ID;
-import static np.com.naxa.factsnepal.common.Constant.JsonKeySurvey.KEY_QN_ID;
-import static np.com.naxa.factsnepal.common.Constant.JsonKeySurvey.KEY_QN_TYPE;
 import static np.com.naxa.factsnepal.common.Constant.JsonKeySurvey.KEY_SURVEY_ID;
 import static np.com.naxa.factsnepal.common.Constant.JsonKeySurvey.KEY_USER_ID;
 import static np.com.naxa.factsnepal.common.Constant.KEY_EXTRA_OBJECT;
@@ -60,11 +50,6 @@ public class SurveyActivity extends BaseActivity {
     private static final String TAG = "SurveyActivity";
 
     Button btnFormSubmit;
-    List<Integer> checkedStringList = new ArrayList<Integer>();
-    String checkboxString = "";
-    String viewTag = "";
-    int viewPosition;
-    String lastViewTag = "";
 
     JSONObject jsonObject = new JSONObject();
     JSONArray jsonArray = new JSONArray();
@@ -72,7 +57,7 @@ public class SurveyActivity extends BaseActivity {
     SurevyForms surevyForms;
     UserLoginResponse userLoginResponse;
 
-    LinearLayout linearLayoutFormList ;
+    LinearLayout linearLayoutFormList;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -91,19 +76,19 @@ public class SurveyActivity extends BaseActivity {
             @Override
             public void onClick(View v) {
                 jsonArray = new JSONArray();
-                getValueFromView();
+                getDataFromJsonVIew();
             }
         });
 
     }
 
-    protected void getNewIntent(Intent intent){
-        if(intent != null){
+    protected void getNewIntent(Intent intent) {
+        if (intent != null) {
             Gson gson = new Gson();
 
-            HashMap<String, Object> hashMap = (HashMap<String, Object>)intent.getSerializableExtra("map");
-            surveyCompany = (SurveyCompany)hashMap.get(KEY_OBJECT);
-            surevyForms = (SurevyForms)hashMap.get(KEY_EXTRA_OBJECT);
+            HashMap<String, Object> hashMap = (HashMap<String, Object>) intent.getSerializableExtra("map");
+            surveyCompany = (SurveyCompany) hashMap.get(KEY_OBJECT);
+            surevyForms = (SurevyForms) hashMap.get(KEY_EXTRA_OBJECT);
             userLoginResponse = gson.fromJson(SharedPreferenceUtils.getInstance(SurveyActivity.this).getStringValue(Constant.SharedPrefKey.KEY_USER_LOGGED_IN_DETAILS, null), UserLoginResponse.class);
 
             try {
@@ -232,223 +217,9 @@ public class SurveyActivity extends BaseActivity {
         }
     }
 
-    private void getValueFromView() {
-        createProgressDialog("Sending data");
-        viewPosition = -1;
-        Log.d(TAG, "getValueFromView: " + Constant.generatedViewIdsList.size());
-        Observable.just(Constant.generatedViewIdsList)
-                .subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .flatMapIterable(new Function<List<Integer>, Iterable<Integer>>() {
-                    @Override
-                    public Iterable<Integer> apply(List<Integer> integers) throws Exception {
-                        return integers;
-                    }
-                })
-                .map(new Function<Integer, Integer>() {
-                    @Override
-                    public Integer apply(Integer integer) throws Exception {
-                        return integer;
-                    }
-                })
-                .subscribe(new DisposableObserver<Integer>() {
-                    @Override
-                    public void onNext(Integer integer) {
-                        Log.d(TAG, "onNext: view id " + integer);
 
-                        try {
-                            View view = findViewById(integer);
-                            viewTag = view.getTag().toString();
-                            viewPosition++;
-                            getChildFromLinearLayout(view);
-                        } catch (Exception e) {
-                            e.printStackTrace();
-                        }
-
-                    }
-
-                    @Override
-                    public void onError(Throwable e) {
-                        hideProgressDialog();
-                        showToast(e.getMessage());
-                        e.printStackTrace();
-                    }
-
-                    @Override
-                    public void onComplete() {
-                        try {
-                            jsonObject.put(Constant.JsonKeySurvey.KEY_ANSWER, jsonArray);
-                            postFormDataToSerever();
-                            Log.d(TAG, "json to send : "+jsonObject.toString());
-                        } catch (JSONException e) {
-                            e.printStackTrace();
-                        }
-
-                    }
-                });
-
-    }
-
-
-    public synchronized void getChildFromLinearLayout(View view) throws JSONException {
-        if (view == null) {
-            Log.d(TAG, "getChildFromLinearLayout: Null View");
-            return;
-        }
-        Log.d(TAG, "getChildFromLinearLayout: " + view.getClass().getSimpleName());
-
-
-        switch (view.getClass().getSimpleName()) {
-            case "TextView":
-                getValueFromTextView((TextView) view);
-                break;
-
-            case "RadioGroup":
-                    jsonArray.put(getValueFromRadioGroup((RadioGroup) view) );
-                Log.d(TAG, "getChildFromRadioGroup: " + ((RadioGroup) view).getChildCount());
-                break;
-
-            case "RatingBar":
-                jsonArray.put(getValueFromRatingBar((RatingBar) view) );
-                break;
-
-            case "Spinner":
-                    jsonArray.put(getValueFromSpinner((Spinner) view) );
-                break;
-
-            case "CheckBox":
-                getValueFromCheckBox((CheckBox) view);
-                if (TextUtils.equals(viewTag, lastViewTag)) {
-                    try {
-                        jsonArray.put(getValueFromCheckBox((CheckBox)view));
-                    } catch (JSONException e) {
-                        e.printStackTrace();
-                    }
-
-                }
-                break;
-
-            case "LinearLayout":
-                LinearLayout linearLayout = (LinearLayout)view ;
-                if(linearLayout.getChildAt(0) instanceof CheckBox){
-                    jsonArray.put(getValueFromCheckedLinearLayout(linearLayout));
-                    Log.d(TAG, "getChildFromLinearLayout: "+linearLayout.getChildCount());
-                }
-                break;
-
-            case "EditText":
-                jsonArray.put(getValueFromEditTex((EditText) view) );
-                break;
-
-        }
-
-    }
-
-    public synchronized JSONObject getValueFromCheckedLinearLayout (LinearLayout linearLayout) throws JSONException{
-        List<Integer> checkedStringList1 = new ArrayList<Integer>();
-        for (int i = 0 ; i<linearLayout.getChildCount(); i++){
-            CheckBox checkBox = (CheckBox)linearLayout.getChildAt(i);
-            if(checkBox.isChecked()){
-                checkedStringList1.add(Integer.parseInt(checkBox.getTag().toString()));
-            }
-        }
-        JSONObject jsonObject = new JSONObject();
-        jsonObject.put(KEY_QN_TYPE, Constant.ViewTypeStringKey.KEY_CHECKBOX);
-        jsonObject.put(KEY_QN_ID, viewTag);
-        jsonObject.put(KEY_ANS_ID, checkedStringList1);
-
-        return jsonObject;
-    }
-
-    public synchronized JSONObject getValueFromRadioGroup(RadioGroup radioGroup) throws JSONException {
-        RadioButton rb1 = (RadioButton) findViewById(radioGroup.getCheckedRadioButtonId());
-        if (rb1 == null) {
-            showToast("no any option selected");
-            throw new  JSONException("no any option selected");
-        }
-//        String radiovalue = rb1.getText().toString();
-        int radiovalue = Integer.parseInt(rb1.getTag().toString());
-        Log.d(TAG, "getValueFromRadioGroup: radiovalue " + radiovalue);
-
-        JSONObject jsonObject = new JSONObject();
-        jsonObject.put(KEY_QN_TYPE, Constant.ViewTypeStringKey.KEY_RADIO);
-        jsonObject.put(KEY_QN_ID, viewTag);
-        jsonObject.put(KEY_ANS_ID, radiovalue);
-        return jsonObject;
-    }
-
-    public synchronized JSONObject getValueFromTextView(TextView textView) throws JSONException {
-        Log.d(TAG, "getValueFromTextView: " + textView.getClass().getSimpleName() + " " + textView.getText().toString());
-        JSONObject jsonObject = new JSONObject();
-        jsonObject.put(KEY_QN_TYPE, Constant.ViewTypeStringKey.KEY_TEXTVIEW);
-        jsonObject.put(KEY_QN_ID, viewTag);
-        jsonObject.put(KEY_ANS_ID, textView.getText().toString());
-        return jsonObject;
-    }
-
-    public synchronized JSONObject getValueFromEditTex(EditText editText) throws JSONException {
-        Log.d(TAG, "getValueFromTextView: " + editText.getClass().getSimpleName() + " " + editText.getText().toString());
-
-        JSONObject jsonObject = new JSONObject();
-        jsonObject.put(KEY_QN_TYPE, Constant.ViewTypeStringKey.KEY_EDITTEXT);
-        jsonObject.put(KEY_QN_ID, viewTag);
-        jsonObject.put(KEY_ANS_ID, editText.getText().toString());
-        return jsonObject;
-    }
-
-    public synchronized JSONObject getValueFromCheckBox(CheckBox checkBox) throws JSONException {
-
-        if (checkBox.isChecked()) {
-
-            if (TextUtils.equals("",viewTag) || TextUtils.equals(lastViewTag,viewTag)) {
-                checkedStringList.add(Integer.parseInt(checkBox.getTag().toString()));
-            } else {
-//                lastViewTag = checkBox.getTag().toString();
-                checkedStringList = new ArrayList<Integer>();
-                checkedStringList.add(Integer.parseInt(checkBox.getTag().toString()));
-            }
-        } else {
-            int position = checkedStringList.indexOf(Integer.parseInt(checkBox.getTag().toString()));
-            checkedStringList.remove(position);
-        }
-
-        lastViewTag = viewTag;
-        checkboxString = checkedStringList.toString();
-        Log.d(TAG, "getValueFromCheckBox: " + checkboxString);
-
-        JSONObject jsonObject = new JSONObject();
-        jsonObject.put(KEY_QN_TYPE, Constant.ViewTypeStringKey.KEY_CHECKBOX);
-        jsonObject.put(KEY_QN_ID, viewTag);
-        jsonObject.put(KEY_ANS_ID, checkedStringList);
-        return jsonObject;
-    }
-
-    public synchronized JSONObject getValueFromRatingBar(RatingBar ratingBar)throws JSONException {
-        String rating = String.valueOf(ratingBar.getRating());
-
-        Log.d(TAG, "getValueFromratingBar: " + rating);
-        JSONObject jsonObject = new JSONObject();
-        jsonObject.put(KEY_QN_TYPE, Constant.ViewTypeStringKey.KEY_RATING_BAR);
-        jsonObject.put(KEY_QN_ID, viewTag);
-        jsonObject.put(KEY_ANS_ID, rating);
-        return jsonObject;
-    }
-
-    public synchronized JSONObject getValueFromSpinner(Spinner spinner) throws JSONException {
-        int selectedSpinnerValuePos = spinner.getSelectedItemPosition();
-        int selectedSpinnerValueID = buildUi().optJSONObject(viewPosition).optJSONArray("options").optJSONObject(selectedSpinnerValuePos).optInt("id");
-        Log.d(TAG, "getValueFromSpinner: TAG "+selectedSpinnerValueID);
-
-        JSONObject jsonObject = new JSONObject();
-        jsonObject.put(KEY_QN_TYPE, Constant.ViewTypeStringKey.KEY_SPINNER);
-        jsonObject.put(KEY_QN_ID, viewTag);
-        jsonObject.put(KEY_ANS_ID, selectedSpinnerValueID);
-        return jsonObject;
-    }
-
-
-    private synchronized void generateViewFromJSON (@NonNull JSONArray jsonArray1){
-        if(jsonArray1 == null){
+    private synchronized void generateViewFromJSON(@NonNull JSONArray jsonArray1) {
+        if (jsonArray1 == null) {
             showToast("null data \ncan't create view");
             return;
         }
@@ -458,7 +229,7 @@ public class SurveyActivity extends BaseActivity {
                 .flatMap(new Function<JSONArray, ObservableSource<JSONObject>>() {
                     @Override
                     public ObservableSource<JSONObject> apply(JSONArray jsonArray) throws Exception {
-                        return Observable.range(0,jsonArray.length())
+                        return Observable.range(0, jsonArray.length())
                                 .map(new Function<Integer, JSONObject>() {
                                     @Override
                                     public JSONObject apply(Integer index) throws Exception {
@@ -467,12 +238,12 @@ public class SurveyActivity extends BaseActivity {
                                 });
                     }
                 })
-               .map(new Function<JSONObject, JSONObject>() {
-                   @Override
-                   public JSONObject apply(JSONObject jsonObject) throws Exception {
-                       return jsonObject;
-                   }
-               })
+                .map(new Function<JSONObject, JSONObject>() {
+                    @Override
+                    public JSONObject apply(JSONObject jsonObject) throws Exception {
+                        return jsonObject;
+                    }
+                })
                 .subscribe(new DisposableObserver<JSONObject>() {
                     @Override
                     public void onNext(JSONObject jsonObject) {
@@ -539,8 +310,32 @@ public class SurveyActivity extends BaseActivity {
     }
 
 
+    private void getDataFromJsonVIew() {
+        createProgressDialog("Sending data");
+        new GetDataFromJsonView(SurveyActivity.this, buildUi()) {
+            @Override
+            public void onErrorListner(String errorMsg) {
+                showToast(errorMsg);
+                hideProgressDialog();
+            }
 
-    private void postFormDataToSerever (){
+            @Override
+            public void onCompleteListner(JSONArray answerJsonArray) {
+                Log.d(TAG, "json onCompleteListner: " + answerJsonArray);
+                try {
+                    jsonObject.put(KEY_ANSWER, answerJsonArray);
+                    postFormDataToSerever();
+                } catch (JSONException e) {
+                    showToast(e.getMessage());
+                    hideProgressDialog();
+                    e.printStackTrace();
+                }
+            }
+        };
+    }
+
+
+    private void postFormDataToSerever() {
         apiInterface.postSurveyAnswerDetailsResponse(jsonObject.toString())
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
@@ -549,7 +344,7 @@ public class SurveyActivity extends BaseActivity {
                     public void onNext(PostSurveyAnswerResponse postSurveyAnswerResponse) {
 
                         hideProgressDialog();
-                        if (postSurveyAnswerResponse.getSuccess()){
+                        if (postSurveyAnswerResponse.getSuccess()) {
                             showToast(postSurveyAnswerResponse.getMessage());
                             Constant.generatedViewIdsList = null;
                             Constant.generatedViewIdsList = new ArrayList<Integer>();
@@ -563,7 +358,7 @@ public class SurveyActivity extends BaseActivity {
                     @Override
                     public void onError(Throwable e) {
                         hideProgressDialog();
-                        if(e instanceof EOFException || e instanceof SocketTimeoutException){
+                        if (e instanceof EOFException || e instanceof SocketTimeoutException) {
                             createProgressDialog("Retrying Data send");
                             postFormDataToSerever();
                         }
